Update patch set 1
Patch Set 1:

> Only stupid thing I see here is how this was handled, making all
 > this drama for something was done 1 year ago, With all due respect,
 > but you honestly remind me my teenager boy who's always shouting
 > why his 8y old sister is looking at him, taking his stuff, going in
 > to his room, drinking his water, etc, etc. I like what my co-worker
 > always tell me which is you have to know how to choose your
 > battles,I can understand that you didn't like the whole thing, I
 > can understand that you probably dont like the guy, but do you
 > really think you get more being this way ? each his own .. Have a
 > good night guys. !

You're right, this was handled poorly. You guys should of just merged it and not said anything. Instead you guys had to run your mouth. It was really stupid how this was handled. Thank you Hector for providing further proof to that by taking this to a personal level and continuing to talk trash, you are a class act!

Patch-set: 1