Update patch set 1
Patch Set 1:

> > Don't see why this comment was removed in the first place now. No
 > > disrespect was made to anyone and the comment was added in the
 > time
 > > why Alex was exposing some rom kangers.
 > >
 > > And stop acting like children all of you God damn it.
 > >
 > > He said, she said... Everyone here is more mature than it shows.
 > >
 > > And then you wonder why I'm not active anymore.. Because of stuff
 > > like this that escalate to high skies with no reason and you lose
 > > good friends for basically nothing.
 > 
 > It wouldn't be a bad idea for you to take your project back. I
 > believe AICP ran smoother, both the ROM and project itself when you
 > were active. These guys are running things like thugs.

I have never dignified your aggressive comments to anyone with an answer. So far. Now you are out of line. It's not your place to make any comments on how an operation is run, who is running it, what they are doing etc. etc. etc. So stay out of it. You don't see us coming to your turf and talk trash about you or any member in your team, do you? So show the same respect. As you said you got what you came here for and it's time for you to leave. Buh bye.

Patch-set: 1